requirements.txt
google_search_results==2.4.2
langchain==0.3.23
langchain_chroma==0.2.3
langchain_community==0.3.21
langchain_core==0.3.54
langchain_groq==0.3.2
langchain_huggingface==0.1.2
python-dotenv==1.1.0
streamlit==1.44.1